Move ApiRestSection to its own directory and rename its style.ts
